+++ pf4j/src/main/java/org/pf4j/PluginClassLoader.java
@@ -181,9 +181,8 @@ public class PluginClassLoader extends URLClassLoader {
      */
     @Override
     public URL getResource(String name) {
-        ClassLoadingStrategy loadingStrategy = getClassLoadingStrategy(name);
         log.trace("Received request to load resource '{}'", name);
-        for (ClassLoadingStrategy.Source classLoadingSource : loadingStrategy.getSources()) {
+        for (ClassLoadingStrategy.Source classLoadingSource : classLoadingStrategy.getSources()) {
             URL url = null;
             switch (classLoadingSource) {
                 case APPLICATION:
@@ -211,9 +210,8 @@ public class PluginClassLoader extends URLClassLoader {
     @Override
     public Enumeration<URL> getResources(String name) throws IOException {
         List<URL> resources = new ArrayList<>();
-        ClassLoadingStrategy loadingStrategy = getClassLoadingStrategy(name);
         log.trace("Received request to load resources '{}'", name);
-        for (ClassLoadingStrategy.Source classLoadingSource : loadingStrategy.getSources()) {
+        for (ClassLoadingStrategy.Source classLoadingSource : classLoadingStrategy.getSources()) {
             switch (classLoadingSource) {
                 case APPLICATION:
                     if (getParent() != null) {
@@ -230,14 +228,6 @@ public class PluginClassLoader extends URLClassLoader {
         }
 
         return Collections.enumeration(resources);
-    }
-
-    private ClassLoadingStrategy getClassLoadingStrategy(String name) {
-        ClassLoadingStrategy loadingStrategy = classLoadingStrategy;
-        if (IndexedExtensionFinder.EXTENSIONS_RESOURCE.equals(name)) {
-            loadingStrategy = ClassLoadingStrategy.PAD;
-        }
-        return loadingStrategy;
     }
 
     /**
@@ -315,14 +305,16 @@ public class PluginClassLoader extends URLClassLoader {
         log.trace("Search in dependencies for resource '{}'", name);
         List<PluginDependency> dependencies = pluginDescriptor.getDependencies();
         for (PluginDependency dependency : dependencies) {
-            PluginClassLoader classLoader = (PluginClassLoader) pluginManager.getPluginClassLoader(dependency.getPluginId());
+            ClassLoader classLoader = pluginManager.getPluginClassLoader(dependency.getPluginId());
 
             // If the dependency is marked as optional, its class loader might not be available.
             if (classLoader == null && dependency.isOptional()) {
                 continue;
             }
 
-            URL url = classLoader.findResource(name);
+            URL url = classLoader instanceof URLClassLoader
+                ? ((URLClassLoader) classLoader).findResource(name)
+                : classLoader.getResource(name);
             if (Objects.nonNull(url)) {
                 return url;
             }
@@ -343,14 +335,16 @@ public class PluginClassLoader extends URLClassLoader {
         List<URL> results = new ArrayList<>();
         List<PluginDependency> dependencies = pluginDescriptor.getDependencies();
         for (PluginDependency dependency : dependencies) {
-            PluginClassLoader classLoader = (PluginClassLoader) pluginManager.getPluginClassLoader(dependency.getPluginId());
+            ClassLoader classLoader = pluginManager.getPluginClassLoader(dependency.getPluginId());
 
             // If the dependency is marked as optional, its class loader might not be available.
             if (classLoader == null && dependency.isOptional()) {
                 continue;
             }
 
-            results.addAll(Collections.list(classLoader.findResources(name)));
+            results.addAll(Collections.list(classLoader instanceof URLClassLoader
+                ? ((URLClassLoader) classLoader).findResources(name)
+                : classLoader.getResources(name)));
         }
 
         return results;