+++ CHANGELOG.md
@@ -7,8 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Fixed
 - [#669]: Read extension annotations from class files produced by recent Java releases
 - [#670]: Read the `ordinal` attribute of the `Extension` annotation via ASM
+- [#673]: Do not report the extensions of a dependency or of the application as extensions of a plugin
+- [#676]: Read the extensions of a plugin loaded with a custom class loader without a `ClassCastException`
 
 #### Changed
+- [#673]: Read the extensions index of a plugin from every jar on the plugin classpath
 
 #### Added
 
@@ -561,6 +564,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [0.11.0]: https://github.com/decebals/pf4j/compare/release-0.10.0...release-0.11.0
 [0.10.0]: https://github.com/decebals/pf4j/compare/release-0.9.0...release-0.10.0
 
+[#676]: https://github.com/pf4j/pf4j/issues/676
+[#673]: https://github.com/pf4j/pf4j/issues/673
 [#670]: https://github.com/pf4j/pf4j/issues/670
 [#669]: https://github.com/pf4j/pf4j/issues/669
 [#656]: https://github.com/pf4j/pf4j/issues/656